Changed Signup process styling and added sidenavbars to few pages
