+++ src/App.scss
@@ -3,10 +3,11 @@
 @import url('https://fonts.googleapis.com/css?family=Inter');
 @import url('https://fonts.googleapis.com/css?family=Josefin+Sans');
 
-.app {
-    display: flex;
+// .app {
+//     display: flex;
+//     justify-content: center;
     
-}
-* {
-    font-family: $title-font-family;
-}
+// }
+// * {
+//     font-family: $title-font-family;
+// }
