+++ src/components/ContactListPage/ContactListPage.jsx
@@ -3,15 +3,14 @@ import ContactContainer from '../../containers/ContactContainer/ContactContainer
 import NavBar from "../../assets/images/navbar.png";
 import SideBar from "../../assets/images/sidebar.jpg";
 import  "./ContactListPage.scss";
-import { Link } from 'react-router-dom';
-
+import SignInNav from '../SignInNav/SignInNav';
+import NavMenu from '../NavMenu/NavMenu';
 
 const ContactListPage = () => {
   return (
-    // <Link to="/contacts">
         <div className="contactPage">
             <div className="placeHolders__left">
-                <img src={SideBar} alt="SideBar" />
+                <NavMenu />
             </div>
             <div className="placeHolders__right">
                 <div className="placeHolders__Navbar">
@@ -20,7 +19,6 @@ const ContactListPage = () => {
                 <ContactContainer />
             </div>
         </div>
-    // </Link> 
   )
 }
