+++ src/components/LiveRates/LiveRates.jsx
@@ -3,15 +3,17 @@ import { useParams } from 'react-router-dom';
 import currency from '../../data/currency.js'
 import LiveRatesList from '../LiveRatesList/LiveRatesList.jsx';
 import "./LiveRates.scss"
+import NavMenu from '../NavMenu/NavMenu.jsx';
 
 
 
 const LiveRates = () => {
     const {liverates} = useParams();
   return (
+  <div className='live-rates-page'>
+<NavMenu />
     
     <div className='live-rates'>
-
         <h4 className='live-rates__h4'>Transfer</h4>
         <h2 className='live-rates__h2'>View Latest Rates</h2>
         <p className='live-rates__para'>Lorem ipsum dolor sit amet consectetur adipisicing elit. Veniam, voluptatem quidem? Corporis, odio vero sequi vitae deleniti sit quod quae! Autem unde reiciendis sunt porro obcaecati alias fugiat quod aliquid!
@@ -20,7 +22,7 @@ const LiveRates = () => {
         <button className='button secondary'>Add Currency</button>
        
     </div>
-
+    </div>
   )
 }
