+++ src/containers/SignUp/SignUpMain.scss
@@ -1,3 +1,5 @@
-.signin__container {
+.signupmain__container {
     margin:5px;
+    display: flex;
+    justify-content: center;
 }
